Windows Additions: Added a new tool VBoxNSISTool for verifying + repairing NSIS installer files, only used at build time. This is necessary for getting the CRC32 right after we've built the NSIS installer and set the version via VBoxSetPeVersion afterwards, which renders the initial checksum within the installer invalid and thus prevents starting the installer [build fix]. bugref:10761
git-svn-id: https://www.virtualbox.org/svn/vbox/trunk@107270 cfe28804-0f27-0410-a406-dd0f0b0b656f

+++ src/VBox/Additions/WINNT/Installer/NSISTool/VBoxNSISTool.cpp
@@ -78,15 +78,15 @@ typedef struct VBOXNSISFILE
     RTFILE      hFile;
     /** File size (in bytes).
      *  Set to 0 if not set (yet). */
-    size_t      cbFile;
+    uint64_t    cbFile;
     /** Header structure, taken from NSIS. */
     firstheader FirstHdr;
     /** Absolute file offset (in bytes) of \a FirstHdr.
      *  Set to UINT64_MAX if not set (yet). */
-    size_t      offHdr;
+    uint64_t    offHdr;
     /** Absolute file offset (in bytes) of CRC32 to read/write.
      *  Set to UINT64_MAX if not set (yet). */
-    size_t      offCRC32;
+    uint64_t    offCRC32;
     /** Currente CRC32 checksum being used.
      *  Set to 0 if not initialized. */
     uint32_t    uCRC32;
